Fix some english grammar Add some clearer docblocks & return types Move integration test to its relevant namespace
